refactor: clean up utils exports
