+++ ai_dubbing/src/audio_processor.py
@@ -11,7 +11,7 @@ import soundfile as sf
 
 # 使用绝对导入
 from ai_dubbing.src.config import AUDIO
-from ai_dubbing.src.utils.common_utils import create_directory_if_needed
+from ai_dubbing.src.utils import create_directory_if_needed
 from ai_dubbing.src.logger import get_logger
