+++ ai_dubbing/src/utils/__init__.py
@@ -4,5 +4,25 @@
 提供项目中常用的工具函数和类
 """
 
-from .common_utils import *
-from ai_dubbing.src.optimizer.subtitle_optimizer import LLMContextOptimizer
+from .common_utils import (
+    setup_project_path,
+    validate_file_exists,
+    create_directory_if_needed,
+    format_duration,
+    format_progress_text,
+    ProgressLogger,
+    normalize_audio_data,
+    initialize_project,
+)
+
+__all__ = [
+    "setup_project_path",
+    "validate_file_exists",
+    "create_directory_if_needed",
+    "format_duration",
+    "format_progress_text",
+    "ProgressLogger",
+    "normalize_audio_data",
+    "initialize_project",
+]
+
