+++ ai_dubbing/validate_durations.py
@@ -15,7 +15,7 @@ project_root = Path(__file__).parent.parent
 sys.path.insert(0, str(project_root))
 
 from ai_dubbing.src.parsers.srt_parser import SRTParser
-from ai_dubbing.src.optimizer.subtitle_optimizer import LLMContextOptimizer, SubtitleTimingConstants
+from ai_dubbing.src.optimizer.subtitle_optimizer import SubtitleTimingConstants
 from ai_dubbing.src.logger import get_logger
